Add short_name helper function to bundle init templates
