+++ internal/testdata/init/helper-short-name/template/foo.txt.tmpl
@@ -1,0 +1,1 @@
+{{short_name}}
