+++ internal/testdata/init/helper-user-name/databricks_template_schema.json
@@ -1,0 +1,1 @@
+{}
