+++ internal/testdata/init/helper-user-name/template/foo.txt.tmpl
@@ -1,0 +1,1 @@
+{{user_name}}
