+++ internal/testdata/init/helper-workspace-host/databricks_template_schema.json
@@ -1,0 +1,1 @@
+{}
