+++ internal/testdata/init/helper-workspace-host/template/foo.txt.tmpl
@@ -1,0 +1,1 @@
+{{workspace_host}}
